// apps/backend/cloudflare-sfu/types.ts
export interface components {
	schemas: {
		SessionDescription: {
			sdp?: string;
			/** @enum {string} */
			type?: "answer" | "offer";
		};
		TrackObject: {
			/**
			 * @description If you want to share a track, it should be local. If you want to play a track shared by a remote agent, it should be remote
			 * @enum {string}
			 */
			location?: "local" | "remote";
			/** @description mid associated to track's transceiver. It also can be prefixed with \# to reference an existing transceiver by its trackName */
			mid?: string;
			/** @description Session ID of the track owner. It should be set for remote tracks only */
			sessionId?: string;
			/** @description Given name for the track */
			trackName?: string;
			/** @description Make the associated transceiver bidirectional. This option works only when the SFU generates the offer */
			bidirectionalMediaStream?: boolean;
			/** @description Give a hint to the SFU about the transceiver kind. This is required when the SFU generates the offer */
			kind?: string;
			/** @description Simulcast configuration for the track */
			simulcast?: {
				/** @description Preferred RID for simulcast streams */
				preferredRid?: string;
				/**
				 * @description Controls what happens if there is not enough network resources available to send the preferredRid. 'none' means keep sending even if not enough bandwidth, 'asciibetical' uses a-z order to determine priority where a is most desirable and z is least desirable.
				 * @default none
				 * @enum {string}
				 */
				priorityOrdering: "none" | "asciibetical";
				/**
				 * @description Controls what happens when the rid currently being used or preferredRid is no longer being sent by the publisher. 'none' means do nothing, 'asciibetical' uses the next available layer after sorting the layers a-z.
				 * @default none
				 * @enum {string}
				 */
				ridNotAvailable: "none" | "asciibetical";
			};
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}